Go migration: implement Transiter service
This commit implements the "Transiter service", which is the public API
of a Transiter instance. The private API will be implemented separately as an
admin service so that it can listen on a private port.

The API will be cleaned up before v0.6 is released; parts of it are sort of
messy right now as they're following the ad hoc Python API.

This commit also includes a script for diffing the new Go based API with the
legacy Python API.