Refactor the public API implementation
I recently discovered pgx's `BeginTxFunc` which handles all of the boilerplate
around transactions. This refactor started because I wanted to use it in the
public API implementation, instead of relying on the session type I had and
its defer function.

With this approach, the endpoint logic must be contained in a function that
is invoked inside `BeginTxFunc`. The result here is that all of the endpoint
logic is now extracted to free functions, and the server invokes these in
a `BeginTxFunc` type style. To make this less toilsome, I upgraded to Go 1.18
and use generics for the function that does this.