Add NYCT stations.csv parser
This commit adds a parser and updater for the NYC Subway's stations.csv file.
This file is used to get "stop headsign rules" which are used to add stop
headsigns to (realtime) trips.

This feature was previously called "direction rules" or "direction name rules"
but was renamed to be consistent with the analagous field in the GTFS static
spec.