Vendorize the Python GTFS Realtime bindings
The end-to-end tests previously used the gtfs-realtime-bindings library to
create GTFS realtime messages. The problem is this library does not pin the
version of the protobuf library it needs. Last night a backwards incompatable
change to that library was pushed to PyPI, resulting in the gtfs library
to break and Transiter E2E tests to fail. This commit vendorized the bindings
and pins the version of the protobuf library to avoid this problem.