Fix zombie service map bug
When the last realtime trip for a route is removed from a GTFS realtime feed,
we need to delete the realtime service map. This was not working due to a very
subtle map construction bug. I've added a pretty nice E2E test to catch this
regression.

Seeing this bug in the Go version is kind of funny because this was a major
problem in the original Python implementation, too. The feed_update table
and corresponding source_pk columns were added partially to solve this problem!
The tricky thing is that without this data, you can't distinguish between
"this feed doesn't update this route" and "this feed does update this route,
but right now there are just no more trips". The source_pk column on trips is
what gives you this: as you delete stale trips for the feed using source_pk,
you can check the routes that have gone.