Public API resources model
In the Transiter API, both in Python and in Go before this commit, many things
like Stops had a collection of distinct, ad-hoc, representations in the API.
The only unfying design behind these was that they would work well for the
realtimerail app.

This commit adds a very clear and understandable resource hierarchy to the
API. Each thing gets two representation - a regular type and a smaller
preview type. The list and get endpoints uniformly return the regular type;
and the smaller type is used when the resource is refered to somewhere else.