Backwards incompatible API changes (1)
I'm planning to merge the Go migration into mainline very soon. Before I do
this I want to make all of the backwards incompatible changes that are in the
TODOs. This way, after the migration is merged, all changes will be backwards
compatible.

This merges the first batch of changes. I was going to wait to do them all at
once, but I need some of these changes on a separate branch.