Simplify realtime service map update
During realtime updates, some service maps may need to be recomputed and
persisted. Before this commit, the service map and realtime update code were
tightly coupled together. The idea of the code was to only recompute the
map when there was a diff. But calculating this diff involved essentially
constructing the old and new map in the realtime update code. This made that
code much more complex for what was essentially a premature optimization.

In this commit the service maps code is split out entirely. When a request
to recompute a service map is made, the code reads from the database to figure
out the trip paths, etc. The main benefit is that the API of the servicemaps
package is now simpler, and the realtime update code is way simpler. Current
con is that we are recomputing service maps more than necessary, but this
can be improved in subsequent work.