Realtime trip updating optimization
This commit adds a pretty significant optimization to the realtime feed
updater. If the data for a trip has not changed, the logic that updates
the trip stop times is skipped. The data not changing is determined by taking
a hash of the trip's data, and comparing it to the hash from the last time
the trip was updated. This has is persisted in the trips table.

Experiments with the NYC subway show that this can result in 25 to 75 percent
of stop time updates being skipped. This is really significant because pprof
profiling of Transiter shows that during a realtime update almost half the time
is spent waiting for Postgres to update trip_stop_time rows.