Correctly mark failed system installs as failed
Currently for async installs, the system is not marked as failed but instead as
installing. The bug is that the transaction function to mark the system
as failed returns the install error, so psql rolls back the transaction. We
instead need to return nil, so that the status change is commited.

This bug manifested as a client hang on a failed install which is pretty bad.

I discovered this while trying out the new MTA buses config that cedarbaum@
wrote. Install of this system fails as expected because of a seperate bug
in the GTFS library that was fixed by cedarbaum@ in
https://github.com/jamespfennell/gtfs/pull/1