More realtime update optimizations (#100)
From the investigation on the bug (#100), the main slowness in the realtime
updater comes from sending stop time updates one-at-a-time to Postgres. With
this commit the updates changed to delete+insert, and both the delete and
insert queries are batched across the whole update. For the NYC buses, this
commit seems to make the updater at least 50% faster.