Remove synchronous system install
Transiter supported a synchronous system install for historical reasons.
Initially all installs were synchronous, which caused issues with request
timeouts, so aync installs were created. The sync code path was kept around
just because.

Now with the Transiter CLI the synchronous install isn't useful, because the
CLI handles polling the async install so the user experience is the same.
The synchronous install introduces some code complexity, particularly around
feed updates and transactions, which would be great to remove. So it seems
best to destroy it.