Use smaller transactions in feed updates
The feed updater in the ported Go code used a single transaction to process
each feed update. The problem is that most (?) of the time in the feed update
is not spent doing database work - it's spent getting the feed data over HTTP
and parsing it. This makes the risk of racing transactions higher. There was
also a bug in which partial updates could be commited because the transaction
was always commited and never rolled back.

With this commit 2 smaller transaction are used to process the update.