Change stale resource deletion impl (#103)
I want to remove the "feed update" concept in Transiter. The immediate
motivation is make system deletion faster (#103) but I think it will make the
software simpler. I will provide more details in the commit that removes
feed updates.

In this commit, I remove one use case for feed updates, which is deleting stale
resources. The model behind both GTFS formats is that each update contains
a full snapshot of the current state of the system. If a resource goes away
(for example, a realtime trip ends) that is not communicated explicitly in
the feed. Software like Transiter needs a way to detect resources
disappearing like this.

The current solution for Transiter is that each resource has a foreign key
reference to the feed update that contained the resource. After updating
a feed, we find all resources that don't foreign key into the most recent
feed update but instead foreign key into an old feed update. Necessarily,
these resources were missing in the most recent update and are then deleted.

This commit changes the implementation of this idea slightly. When performing
an update we keep track of the primary keys of all resources that were updated.
At the end we delete all resources corresponding to a feed whose primary key
is not in this set.

Logically, this gives the same result. The benefit though is that it doesn't
rely on the feed update concept at all. The one downside is that the update
code is a tiny bit more complicated because we need to keep track of the
updated resources' primary keys. However I think this is worth it in light
of the simplification removing feed updates will give us.