Fix disappearing trips bug (#106)
In https://github.com/jamespfennell/transiter/commit/72ee363c1d2de752f004f4d540322612cb2a45ee
I changed the logic for stale entity deletion. There was a pretty bad bug
in this change in which many active trips are considered stale and are deleted.

The problem occurs for trips whose GTFS data has not changed. SQL updates for
these trips are skipped as an optimization. However when skipping such trips
I didn't add them to the set of "active trips", and thus at the end of
the update such trips are deleted as stale.

There is a unit test that explicitly targets this case, but there was bug in
the unit test such that it didn't cover the case correctly. Additionally
the end to end tests have test cases that are close to covering this scenario,
but the larger logic of "skip the entire update if the bytes are the same"
kicks in and again the case is not covered. I have fixed both the unit test
and the end to end tests so that they fail with this bug.