Support updating feeds daily (#109)
Transiter has always supported updating feeds periodically, with an expectation
that the period will be small (default 5 seconds). This is a good choice for
GTFS realtime feeds, but generally a poor choice for GTFS static feeds.
Even with a large period (say 1 day) the time when the update happens is
literally psuedorandom. These updates may be heavy and cause degraded
performance.

This commit introduces a new "scheduling policy" (formally called an "update
strategy") in which a feed is updated once per day at a specified time in a
specified timezone. The nicest part of this work IMO is that the defaults
are very nice: by default Transiter will update a daily feed during the night
in the timezone of the feed. The timezone is determined by looking at the
agency table in the GTFS static data itself.

This commit also changes the default scheduling policies. Before, the default
was no updates via the scheduler. Now, GTFS realtime feeds will by default
be updated once every 5 seconds and all other feeds will updated once a day.
This removes some configuration burden.

I also took the opportunity to simply the feed configuration in a few other
ways:

- Parser is now called type. I think this better aligns with how people think
  about feeds - it's a GTFS realtime feed, not a feed that is parsed using
  the GTFS realtime parser.

- The periodic update period is specified in integer milliseconds.

- The required_for_install field is optional, with a sensible default depending
  on the feed type.

With these changes, adding a new feed is really easy: generally you just
need to specify a feed ID, the feed type, and a URL and Transiter will do the
right thing.

These changes have also been designed to be backwards compatible. This is
tricky but I think I've done it correctly.