Fix integer overflow bug in stops pagination
If a negative number is specified for `MaxStopsPerRequest`, this
allows an unlimited number of stops to be returned by setting the
`numStops` limit to `MaxInt32`. However, the pagination logic further
increases this by one to over-fetch and return the next stop's
ID. This causes overflow and the request to fail.

Simply setting `numStops` to `MaxInt32 - 1` fixes this bug while still
effectively allowing an unlimited stops to be returned.