Add just command runner
I recently learned about just, which is kind of like make but optimized for
the usecase where your makefile is just a collection of commands rather than
actual build steps. I think this is a great addition to the Transiter repo
because it allows us to centralize a lot of command (from the CI config,
random bash scripts, Dockerfiles) into one place.

In addition I take the opportunity to simplify the CI and also update some
of the docs.