Use PostGIS (#118)
Transiter supports quering stop and vehicle data using lat/long coordinates.
Originally this was done using homegrown queries to implement geographic
search. With this commit the project migrates to using the PostGIS extension
in Postgres.

The main con is that Transiter will always require PostGIS to be available.
This isn't too bad though as PostGIS is one of the standard Postgres
extensions.

The pros are simpler and faster queries. It also unlocks implementating
fancier geographic queries in the future, like finding nearby routes.