Easier transit system installs (#115)
This commit makes it possible to install a transit system by just providing
an ID like us-ca-bart. The system config will be automatically pulled from
the Transiter GitHub repo.

It is still possible to install a system by providing a local file or a URL.

Additionally, this commit changes how URL-based installs work. Previously
the URL was provided to the Transiter server which made a GET request to get
the config. This design is a bit wierd from an API perspective, and is an
artifact of the Python version of Transiter which didn't have a client. After
this commit the install system API requires the YAML config itself. For URL
installs, the transiter client performs the GET request.