E2E tests: add a Transiter client
The E2E tests are a really valuable part of the project and are
the main reason we can have confidence that Transiter works correctly when
we make changes. However, the implementation of the E2E tests does not bring
me joy. Historically they been mostly hacked together and rely on lots of
untyped and potentially fragile JSON blob manipulation.

This presents a somewhat small risk to the overall Transiter project. If the
tests are hard to maintain and/or extend, they won't be maintained/extended
and their value over time will go down as they cover fewer features.

This commit starts implementing an idea which I think could make the tests
really pleasant to work with. The idea is to have a typed Transiter client
through which all API interactions occur. With this approach all the JSON blobs
get replaced with Python dataclasses. Instead of checking the JSON blob or
some subset of its fields, we check the dataclass value entirely.

This commit adds the first version of the client and uses it in the test
that verifies the stops endpoints. I also extract the stops test into its own
file. I also run the black autoformatter, which accounts for most of the
diffs.