E2E tests: framework for data independent tests
From the very start the 200+ E2E tests have shared test data in the form
a global GTFS static file. This has been problematic for a few reasons:

- When the global data needs to updated for a new test, it sometimes requires
  other tests to be updated.

- There is a gap between the setup and expectations of the test. I.e., to
  assert the correct route data is returned you need to consult the global
  static data, rather than some setup within the test.

This commit adds a new framework to support data independence in tests. The
idea is simple: every E2E test should define its own GTFS static data that
it relies on. The technical thing that makes this not super tedious is Russ
Cox's txtar format [1]. With this format we can provide an entire GTFS static
data dump in a single string.

One detail is that the GTFS static spec requires some files to be present
(like stops.txt), and even requires agency.txt to have at least 1 row. To
avoid duplicating the default empty file for e.g. stops.txt, this commit adds
a mechanism in which is some GTFS static files are not provided, a working
default is used instead.

The commit implements all of the above and migrates the route and agency
test to use this new framework.

[1] https://research.swtch.com/testing