Fix bug in list stops endpoint
When listing stops all of the lat/lons returned are the same. This is because
of the classic Go for loop footgun. The lat/lons are returned in the API as
pointers into the DB stops, but the pointer is to the loop variable not the
specific value in the current iteration.

I discovered this while migrating the stops E2E tests to use the new txtar
based system install mechanism. I suspect there are other similar bugs in the
API because this kind of test coverage is now.

This commit also improves the stop tests through parameterization, and adds
a test for pagination. This pagination test would have been very very hard
to write without the new txtar mechanism; with txtar it's easy.